Add hero banner

+++ src/components/hero/Hero.tsx
@@ -6,8 +6,8 @@ import "../summer-reception/Weekend.css"
 import { CardProps, getCountryDbName } from "../../global/Global";
 import { toast } from 'react-toastify';
 
-interface GalleryProps extends CardProps{
-    images: string[]; // Array of image objects with unique id and URL
+interface HeroBannerProps extends CardProps {
+    banner: string; // Array of image objects with unique id and URL
 }
 
 interface UploadedImage {
@@ -16,23 +16,23 @@ interface UploadedImage {
     dbUrl: string
 }
 
-const Gallery: React.FC<GalleryProps> = ({ images, country, handleSave, handleDelete, handleCancel, handleBack }) => {
-    const [imagesData, setImagesData] = useState<string[]>([])
-    const [selectedImage, setSelectedImage] = useState<string | null>(null);
+const Hero: React.FC<HeroBannerProps> = ({ banner, country, handleSave, handleDelete, handleCancel, handleBack }) => {
+    const [image, setImage] = useState<string>("")
+    const [viewImage, setViewImage] = useState<string | null>(null)
     const [isChanged, setIsChanged] = useState(false)
-    const [imagesToUpload, setImagesToUpload] = useState<UploadedImage[]>([])
-    const [imagesToDelete, setImagesToDelete] = useState<string[]>([])
+    const [imageToUpload, setImageToUpload] = useState<UploadedImage>()
+    const [imageToDelete, setImageToDelete] = useState<string>("")
 
     useEffect(() => {
-        setImagesData(images)
-    }, [images])
+        setImage(banner)
+    }, [banner])
 
     const handleImageClick = (url: string) => {
-        setSelectedImage(url);
+        setViewImage(url);
     };
 
     const closeModal = () => {
-        setSelectedImage(null);
+        setViewImage(null);
     };
 
     const handleUpload = async (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -43,20 +43,20 @@ const Gallery: React.FC<GalleryProps> = ({ images, country, handleSave, handleDe
                 url: URL.createObjectURL(file),
                 dbUrl: `https://firebasestorage.googleapis.com/v0/b/iaeste-ap.appspot.com/o/${country}%2Fgallery%2F${file.name}?alt=media`
             }
-            setImagesData([...imagesData, newImage.url])
-            setImagesToUpload([...imagesToUpload, newImage])
+            setImage(newImage.url)
+            setImageToUpload(newImage)
             setIsChanged(true)
         } else toast.error("Error while uploading file!")
     };
 
     const uploadImagesToStorage = async () => {
         try {
-            const countryName = getCountryDbName(country)
-            imagesToUpload.forEach(async (image: UploadedImage) => {
-                const storageRef = ref(storage, `${countryName}/gallery/${image.file.name}`);
+            if (imageToUpload) {
+                const countryName = getCountryDbName(country)
+                const storageRef = ref(storage, `${countryName}/gallery/${imageToUpload.file.name}`);
                 // Upload to storage
-                await uploadBytes(storageRef, image.file);
-            })
+                await uploadBytes(storageRef, imageToUpload.file);
+            } else toast.error("No image to upload!")
         } catch (error) {
             toast.error("Error uploading files! " + error)
             return null
@@ -65,12 +65,12 @@ const Gallery: React.FC<GalleryProps> = ({ images, country, handleSave, handleDe
 
     const removeImagesFromStorage = async () => {
         try {
-            imagesToDelete.forEach(async (image: string) => {
-                const fileRef = ref(storage, image)
+            if (imageToDelete !== "") {
+                const fileRef = ref(storage, imageToDelete)
                 await deleteObject(fileRef)
-            })
-        } catch(error) {
-            toast.error("Error removing files from storage! " + error)
+            } else toast.error("No image to delete!")
+        } catch (error) {
+            toast.error("Error removing file from storage! " + error)
             return null
         }
     }
@@ -78,31 +78,29 @@ const Gallery: React.FC<GalleryProps> = ({ images, country, handleSave, handleDe
     const onSave = async () => {
         // upload files to storage
         await uploadImagesToStorage()
-        const newGallery = imagesData.map((image: string) => {
-            const newImageExists = imagesToUpload.find(newImage => newImage.url === image)
-            return newImageExists?.dbUrl ?? image
-        })
         // delete files from storage
         await removeImagesFromStorage()
-        handleSave(country, newGallery, "gallery", "Gallery", setIsChanged)
+        // save image
+        // handleSave(country, image, "banner", "Hero banner", setIsChanged)
     }
     const onCancel = async () => {
-        const confimation = await handleCancel(isChanged, setImagesData, images, setIsChanged)
-        if (confimation) {
-            setImagesToDelete([])
-            setImagesToUpload([])
+        const confirmation = await handleCancel(isChanged, setImage, banner, setIsChanged)
+        if (confirmation) {
+            setImageToDelete("")
+            setImageToUpload(undefined)
         }
     }
-    const onDelete = async (index: number, url: string) => {
-        const del = await handleDelete(index, setImagesData, imagesData, setIsChanged)
-        if (del) setImagesToDelete([...imagesToDelete, url])
+    const onDelete = (url: string) => {
+        setImage("")
+        if (url === banner) setImageToDelete(url)
+        setIsChanged(true)
     }
-    const onBack = () => handleBack(isChanged, setImagesData, images, setIsChanged)
+    const onBack = () => handleBack(isChanged, setImage, banner, setIsChanged)
 
     return (
         <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4 table-margins">
-            {imagesData.map((image, index) => (
-                <div key={index} className="relative group max-h-72">
+            {image ?
+                <div className="relative group max-h-72">
                     <img
                         src={image}
                         alt="Gallery item"
@@ -111,7 +109,7 @@ const Gallery: React.FC<GalleryProps> = ({ images, country, handleSave, handleDe
                     />
                     <button
                         className="absolute top-2 right-2 opacity-0 group-hover:opacity-100 transition duration-300 z-10"
-                        onClick={() => onDelete(index, image)}
+                        onClick={() => onDelete(image)}
                     >
                         <i className="fa-solid fa-trash-can text-red-400 hover:text-red-500 text-2xl" />
                     </button>
@@ -119,25 +117,26 @@ const Gallery: React.FC<GalleryProps> = ({ images, country, handleSave, handleDe
                         onClick={() => handleImageClick(image)}>
                         <i className="fa fa-eye text-white text-4xl" />
                     </div>
+                </div> :
+                <div className="relative group flex items-center justify-center border-dashed border-2 border-gray-300 rounded-lg cursor-pointer hover:border-blue-400 h-72">
+                    <input
+                        type="file"
+                        onChange={handleUpload}
+                        className="absolute inset-0 opacity-0 cursor-pointer"
+                        title="Upload image"
+                    />
+                    <i className="fa-solid fa-plus text-gray-400 text-4xl group-hover:text-blue-400"></i>
                 </div>
-            ))}
-            
+            }
+
             {/* Upload Button Card */}
-            <div className="relative group flex items-center justify-center border-dashed border-2 border-gray-300 rounded-lg cursor-pointer hover:border-blue-400 h-72">
-                <input
-                    type="file"
-                    onChange={handleUpload}
-                    className="absolute inset-0 opacity-0 cursor-pointer"
-                    title="Upload image"
-                />
-                <i className="fa-solid fa-plus text-gray-400 text-4xl group-hover:text-blue-400"></i>
-            </div>
+
 
             {/* Popup Modal for Enlarged Image */}
-            {selectedImage && (
+            {viewImage &&
                 <div className="overlay">
                     <div className="relative bg-black rounded-xl shadow-md w-4/5 h-4/5 flex justify-center">
-                        <img src={selectedImage} alt="Enlarged" className="" />
+                        <img src={viewImage} alt="Enlarged" className="" />
                         <button
                             onClick={closeModal}
                             className="absolute top-2 right-2 text-red-500 rounded-full p-1"
@@ -146,11 +145,11 @@ const Gallery: React.FC<GalleryProps> = ({ images, country, handleSave, handleDe
                         </button>
                     </div>
                 </div>
-            )}
+            }
 
-            <CardFooter isChanged={isChanged} onCancel={onCancel} onSave={onSave} onBack={onBack}/>
+            <CardFooter isChanged={isChanged} onCancel={onCancel} onSave={onSave} onBack={onBack} />
         </div>
     );
 };
 
-export default Gallery;
+export default Hero;
